requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.3
gpiozero==1.5.1
Paste==3.4.3
waitress==1.4.4
RPi.GPIO==0.7.0
python-dotenv==0.16.0